updated on kinemaic1

+++ elephant_can/elephant_can/can_teleop.py
@@ -9,6 +9,8 @@ from std_msgs.msg import Bool
 from sensor_msgs.msg import Joy
 from geometry_msgs.msg import Vector3
 from std_msgs.msg import Float32
+from std_msgs.msg import Int32
+from std_msgs.msg import Int8
 
 gain = 2
 
@@ -28,7 +30,7 @@ class ros_node(Node):
         self.velocity_pub = self.create_publisher(Float32MultiArray, 'pub_speed', 10)
         self.joy_pos_pub = self.create_publisher(Vector3, 'joy_position', 10)
 
-        self.shooter_speed_pub = self.create_publisher(Float32, 'shooter', 10)
+        self.shooter_speed_pub = self.create_publisher(Int32, 'shooter', 10)
         self.velocity_timer = self.create_timer(0.01, self.velocity_callback)
         ##
         self.control_type_pub = self.create_publisher(Bool, 'Controller_state', 10)
@@ -41,7 +43,7 @@ class ros_node(Node):
         self.vy = joy_msg.axes[0]
         self.omega = -1 * joy_msg.axes[3]
 
-        shoot_msg = Float32()
+        shoot_msg = Int32()
 
         if (joy_msg.buttons[2] == 1):
             self.store_speed = (float)(self.kinematic.map((float)(joy_msg.axes[2]),1.0,-1.0,0.0,-1000.0))
